Remodel COSE_Key_Ref as COSE_Sign_Args
The "reference to a key" model is not a good fit for all algorithms. For
example, a "Split BBS" algorithm proposed by Cordian Daniluk and Anja Lehmann
for digital identity use cases requires additional parameters that have nothing
to do with the signing key, only with the application context for the signing
operation.